refactor(http): organize interfaces and services by domain
Interfaces/ and Services/ gain one folder per domain (Endpoints, Auth,
Images, Maps, Ultima, Hosting), namespaces aligned to paths per the
code convention, and the service tests mirror the production layout
under Http/Services.
