+++ src/Moongate.Http.Plugin/Endpoints/Accounts/AccountEndpoints.cs
@@ -3,8 +3,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moongate.Core.Interfaces;
 using Moongate.Core.Types;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
+using Moongate.Http.Plugin.Services.Hosting;
 using Moongate.Persistence.Entities;
 using Moongate.Http.Plugin.Data.Api.Accounts;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
