+++ src/Moongate.Http.Plugin/Endpoints/Auth/AuthEndpoints.cs
@@ -1,7 +1,8 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
-using Moongate.Http.Plugin.Interfaces;
+using Moongate.Http.Plugin.Interfaces.Auth;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
 using Moongate.Http.Plugin.Data.Api.Auth;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
 using Serilog;
