+++ src/Moongate.Http.Plugin/Endpoints/Characters/CharacterAdminEndpoints.cs
@@ -2,8 +2,8 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moongate.Http.Plugin.Data;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
+using Moongate.Http.Plugin.Services.Hosting;
 using Moongate.Http.Plugin.Data.Api.Characters;
 using Moongate.Server.Abstractions.Interfaces.Accounts;
