+++ src/Moongate.Http.Plugin/Endpoints/Maps/MapImageEndpoints.cs
@@ -2,7 +2,8 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moongate.Http.Plugin.Data;
-using Moongate.Http.Plugin.Interfaces;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
+using Moongate.Http.Plugin.Interfaces.Maps;
 using Moongate.UO.Data.Types;
 
 namespace Moongate.Http.Plugin.Endpoints.Maps;
