+++ src/Moongate.Http.Plugin/Endpoints/Version/VersionEndpoints.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
-using Moongate.Http.Plugin.Interfaces;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
 using Moongate.Http.Plugin.Data.Api.Version;
 using Moongate.Server.Abstractions.Data.Config;
 using SquidStd.Core.Utils;
