+++ tests/Moongate.Tests/Http/Endpoints/Images/ItemImageAdminEndpointsTests.cs
@@ -3,8 +3,10 @@ using System.Net.Http.Json;
 using DryIoc;
 using Moongate.Http.Plugin.Data;
 using Moongate.Http.Plugin.Endpoints.Images;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
+using Moongate.Http.Plugin.Interfaces.Images;
+using Moongate.Http.Plugin.Interfaces.Ultima;
+using Moongate.Http.Plugin.Services.Images;
+using Moongate.Http.Plugin.Services.Ultima;
 using Moongate.Http.Plugin.Types;
 using Moongate.Tests.Support;
 using Moongate.Ultima.Catalog;
