+++ tests/Moongate.Tests/Http/Endpoints/Maps/MapImageEndpointsTests.cs
@@ -3,8 +3,10 @@ using System.Net.Http.Json;
 using DryIoc;
 using Moongate.Http.Plugin.Data;
 using Moongate.Http.Plugin.Endpoints.Maps;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
+using Moongate.Http.Plugin.Interfaces.Maps;
+using Moongate.Http.Plugin.Interfaces.Ultima;
+using Moongate.Http.Plugin.Services.Maps;
+using Moongate.Http.Plugin.Services.Ultima;
 using Moongate.Tests.Support;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
