+++ tests/Moongate.Tests/Http/Services/Hosting/HttpServerServiceTests.cs
@@ -4,12 +4,12 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moongate.Http.Plugin.Data.Config;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
+using Moongate.Http.Plugin.Services.Hosting;
 using Moongate.Tests.Support;
 using SquidStd.Core.Interfaces.Config;
 
-namespace Moongate.Tests.Http;
+namespace Moongate.Tests.Http.Services.Hosting;
 
 public class HttpServerServiceTests
 {
