+++ tests/Moongate.Tests/Support/TestApiServer.cs
@@ -5,8 +5,10 @@ using Moongate.Core.Interfaces;
 using Moongate.Core.Types;
 using Moongate.Http.Plugin.Data;
 using Moongate.Http.Plugin.Data.Config;
-using Moongate.Http.Plugin.Interfaces;
-using Moongate.Http.Plugin.Services;
+using Moongate.Http.Plugin.Interfaces.Auth;
+using Moongate.Http.Plugin.Interfaces.Endpoints;
+using Moongate.Http.Plugin.Services.Auth;
+using Moongate.Http.Plugin.Services.Hosting;
 using Moongate.Server.Abstractions.Data.Config;
 using Moongate.Http.Plugin.Endpoints.Accounts;
 using Moongate.Http.Plugin.Endpoints.Admin;
